Add drone movement, thread creation, and utils
Added drone movement functionality which allows a drone to move to a random position, enhancing the drone constructor to create threads for each drone instance, and introducing utils to generate random floats. Multiple updates for classes and attributes have been executed to support these additions, and the cleanup of unwanted lines is also done in the main.cpp file. The adjustments moreover provide structure for future improvements such as formatting the Drone constructor better and changing the status to enum data type.