Refactor DroneZone and DroneManager classes
The DroneZone class was refactored into the DroneManager file and is now instantiated by the DroneManager. This change was made to simplify drone management and to facilitate the calculation and assignment of zones. A DroneZone is now created for every drone zone calculated, with accompanying changes in DroneManager's methods. Unnecessary comments were also removed for better legibility.