Implemented tick-based simulation
Added global simulation settings and uses them to implement simulation loops in DroneControl and Drone. It also introduces a tick counter in these two classes to track the number of ticks transpired during the simulation. Starting and stopping of the simulation are now managed by setting a global 'sim_running' flag in Redis.